implement basic CRUD endpoints for all entities.

+++ src/main/java/com/sitco/api/dtos/RegisterUserRequest.java
@@ -8,5 +8,5 @@ public class RegisterUserRequest {
     private String lastName;
     private String email;
     private String password;
-    private Integer roleId;
+    private Byte roleId;
 }
