+++ src/main/java/com/sitco/api/repositories/MoistureTypeRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface MoistureTypeRepository extends CrudRepository<MoistureType, Integer> {
+public interface MoistureTypeRepository extends CrudRepository<MoistureType, Byte> {
     List<MoistureType> findAll();
 }
